Added logging & refactor
