+++ .gitignore
@@ -1,1 +1,2 @@
 project-0
+log
