+++ README.md
@@ -1,3 +1,27 @@
-# project-0 by FelicianoE
-# TempCli
-TempCli is a command line interface tool to retrieve temperatures from cities around the globe.
+# project-0 by Emilio Feliciano aka FelicianoEJ
+Project 0 is command line tool used to retrieve temperatures from cities around the globe through the use of web api.
+
+**How to use:**
+
+Write `project-0` followed by the city name of your choosing.
+
+For example:
+
+```
+project-0 Miami
+```
+
+Expected output:
+> Fetching...
+
+> Temperature for Miami is 73.8 degrees Fahrenheit.
+
+**Flags:**
+
+`-e` is used to export the weather data into a json file.
+
+Example:
+
+```
+project-0 -e Miami
+```
